requirements.txt
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
aiohttp==3.8.6
openai==1.3.0
# Для Yandex GPT (если решите его использовать)
# yandexcloud==0.227.0 